Comment and Refactor a few files
